optimise directory structure
