+++ model/customer.go
@@ -1,0 +1,8 @@
+package model
+
+type Customer struct {
+	ID      int    `json:"id"`
+	Name    string `json:"name"`
+	Address string `json:"address"`
+	PhoneNo int    `json:"phone_no"`
+}
